Rename folder resonator -> resonators
